File dialog added to entry
Win32/Linux version is from texturev. OSX uses native dialog instead of zenity.
win32 has a fix to support win32 target:  OPENFILENAMEA customData is uintptr_t instead of uint64_t.
